refactor: 훅 네이밍 중 불필요한 부분 제거
